+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 - TT-1828 - Increasing size of address preview field in registration journey
 - TT-1199 - Fixing accessibility issues
 - TT-2025 - Update success page content
+- No ticket - Removed old feature flagged items
+- TT-1608-member-request-admin-2
+- TT-1608 - Member To Request Admin
+- No ticket - upgrade django
+
 
 ### Fixed bugs:
 - TT-1817 - Use correct input types for number fields
@@ -18,6 +23,18 @@
 - TT-2089 - fix-leading-0-verification-code
 - TT-1727 - Not prompt collaborators to change business type
 - TT-2096 - Fix layout problem on enrolment of individual that already has sso profile
+- TT-1708 - link user to ingress url on success page
+- TT-2095 - Added "remove from business profile" feature for non-admins
+- TT-1855 - Fix incorrect company type label on enrolment
+- TT-1132 - Improve contrast of breadcrumbs and feedback link
+- TT-1132 - Upgrade directory components to fix accessibility bug
+- no ticket - upgrade pillow
+
+## [2019.11.20](https://github.com/uktrade/directory-sso-profile/releases/tag/2019.11.20)
+[Full Changelog](https://github.com/uktrade/directory-sso-profile/compare/2019.10.30...2019.11.20)
+
+## Hotfix
+- TT-2176 - Fix password form remembering old bad password
 
 ## [2019.10.30](https://github.com/uktrade/directory-sso-profile/releases/tag/2019.10.30)
 [Full Changelog](https://github.com/uktrade/directory-sso-profile/compare/2019.08.14_1...2019.10.30)
@@ -170,3 +187,4 @@
 - TT-1415 - Allow products and services to be cleared
 [directory-constants]: https://github.com/uktrade/directory-constants
 - TT-1417 - Populate "Other" products and services with keywords
+